Adding TestCases and security to main branch.

+++ myappvs/src/main/java/com/CourseManagementSystem/myappvs/event/EventController.java
@@ -35,18 +35,17 @@ public class EventController {
     // Update an event
     @PutMapping("/{id}")
     public ResponseEntity<?> updateEvent(@PathVariable String id, @Validated @RequestBody Event eventDetails) {
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String authenticatedUserEmail = userDetails.getUsername();
+
         Event event = eventRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("Event not found"));
 
-        /*
-         * if (!event.getEmailId().equals(emailId)) {
-         * System.out.println(emailId + "URL Email");
-         * System.out.println(event.getEmailId());
-         * return ResponseEntity.status(HttpStatus.FORBIDDEN)
-         * .body("You are not authorized to update this event. This event is not hosted by you."
-         * );
-         * }
-         */
+        // Check if the authenticated user is the owner of the event
+        if (!event.getEmailId().equals(authenticatedUserEmail)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("You are not authorized to update this event. This event is not hosted by you.");
+        }
 
         event.setDate(eventDetails.getDate());
         event.setEventName(eventDetails.getEventName());
@@ -58,8 +57,17 @@ public class EventController {
     // Delete an event
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteEvent(@PathVariable String id) {
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String authenticatedUserEmail = userDetails.getUsername();
+
         Event event = eventRepository.findById(id)
-                .orElseThrow();
+                .orElseThrow(() -> new IllegalArgumentException("Event not found"));
+
+        // Check if the authenticated user is the owner of the event
+        if (!event.getEmailId().equals(authenticatedUserEmail)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("You are not authorized to delete this event. This event is not hosted by you.");
+        }
 
         eventRepository.delete(event);
         return ResponseEntity.ok().build();
